requirements.txt
Flask==2.2.3
openai==0.27.0
python-dotenv==0.21.0
transformers==4.30.0
torch==2.4.1